Add JSON Web Tokens (JWT) to the API Authentication Layer (#4)
* Fixed merge conflict

* Added the apiController base class

* Updated the loginController to use JWTs

* Added middleware to detect an authorized user and instantiates permissions for them; beginnings of permissions system

* Added babel-cli to the package file for tests

* Added CORS support to the Express app

* Fixed minor errors in userPermissionsProxy

* Added permissionsLevels lib file; removed created_by_id field from University model; removed universityId from User model and fixed use of wrong ApiError.

* Added significant constraints to the Membership model and strong error handling.

* Added Membership creation in a University to the signup route; fixed tests for login and signup

* Fixed slight errors in the userPermissionsProxy constructor university permissions

* Allow for a University to be created at the same time as a new user. Complicated shit...

* Document CORS

+++ test/controllers/loginController.spec.js
@@ -1,11 +1,13 @@
 let mocha = require('mocha')
 let chai = require('chai')
 let chaiHttp = require('chai-http')
+let jwt = require('jwt-simple')
 let _ = require('lodash')
 
 let app = require('../../src/app.js')
 let Sequelize = require('sequelize')
 let db = require('../../src/db.js')
+let config = require('../../config/secrets')
 
 let should = chai.should
 let expect = chai.expect
@@ -40,12 +42,11 @@ describe ('Login Controller', () => {
       }
 
       chai.request(app)
-        .post('/api/login')
+        .post('/api/auth/login')
         .send(payload)
         .end((err, res) => {
           expect(res).to.have.status(200)
-          expect(res.body.state).to.be.eql(true)
-          expect(res.body.user.id).to.be.eql(testUser.id)
+          expect(res.body.token).to.exist
           done()
         })
     })
@@ -57,7 +58,7 @@ describe ('Login Controller', () => {
       }
 
       chai.request(app)
-        .post('/api/login')
+        .post('/api/auth/login')
         .send(payload)
         .end((err, res) => {
           expect(res).to.have.status(401)
@@ -75,12 +76,14 @@ describe ('Login Controller', () => {
       Promise.all(promises).then(() => {
         done()
       })
-    });
+    })
   })
 
   describe ('loginController#postSignup', () => {
     let testUser0 = null
     let testUser1 = null
+    let testUser2 = null
+    let testUni0 = null
 
     beforeEach((done) => {
       let promises = _.concat([],
@@ -102,23 +105,77 @@ describe ('Login Controller', () => {
 
     it ('successfully signs up a new user', (done) => {
       let payload = {
-        email: 'tnpetresky+1@gmail.com',
+        email: 'tnpetresky+test@gmail.com',
         password: 'password',
-        universityId: 1
+        universityId: 1,
+        permissionLevel: 1
       }
 
       chai.request(app)
-        .post('/api/signup')
+        .post('/api/auth/signup')
         .send(payload)
         .end((err, res) => {
           expect(res).to.have.status(200)
-          expect(res.body.state).to.be.eql(true)
+          expect(res.body.token).to.exist
           expect(res.body.user.email).to.be.eql(payload.email)
-          expect(res.body.user.universityId).to.be.eql(payload.universityId)
 
           testUser1 = res.body.user
 
-          done()
+          db.Membership.count({
+            where: {
+              userId: res.body.user.id,
+              universityId: payload.universityId,
+              rsoId: null,
+              active: true
+            }
+          }).then((count) => {
+            expect(count).to.be.eql(1)
+            done()
+          })
+        })
+    })
+
+    it ('successfully signs up a new user and creates a new university', (done) => {
+      let payload = {
+        email: 'tnpetresky+test@gmail.com',
+        password: 'password',
+        permissionLevel: 1,
+        universityName: 'Test University',
+        description: 'My new uni'
+      }
+
+      chai.request(app)
+        .post('/api/auth/signup')
+        .send(payload)
+        .end((err, res) => {
+          expect(res).to.have.status(200)
+          expect(res.body.token).to.exist
+          expect(res.body.user.email).to.be.eql(payload.email)
+
+          testUser1 = res.body.user
+
+          db.University.find({
+            where: {
+              name: payload.universityName
+            }
+          }).then((uni) => {
+            expect(uni).to.not.be.null
+
+            testUni0 = uni
+
+            db.Membership.count({
+              where: {
+                userId: res.body.user.id,
+                universityId: uni.id,
+                rsoId: null,
+                active: true
+              }
+            }).then((count) => {
+              expect(count).to.be.eql(1)
+              done()
+            })
+          })
+
         })
     })
 
@@ -126,11 +183,12 @@ describe ('Login Controller', () => {
       let payload = {
         email: 'tnpetresky+0@gmail.com',
         password: 'password',
-        universityId: 1
+        universityId: 1,
+        permissionLevel: 1
       }
 
       chai.request(app)
-        .post('/api/signup')
+        .post('/api/auth/signup')
         .send(payload)
         .end((err, res) => {
           expect(res).to.have.status(400)
@@ -139,16 +197,43 @@ describe ('Login Controller', () => {
         })
     })
 
+    it ('does not allow users to signup as an ADMIN in a university', (done) => {
+      let payload = {
+        email: 'tnpetresky+test@gmail.com',
+        password: 'password',
+        universityId: 1,
+        permissionLevel: 2
+      }
+
+      chai.request(app)
+        .post('/api/auth/signup')
+        .send(payload)
+        .end((err, res) => {
+          expect(res).to.have.status(400)
+          expect(res.body.errorCode).to.be.eql(402)
+          done()
+        })
+    })
+
     // Drop all of the test objects from the database
     afterEach((done) => {
       let promises = _.concat([],
         db.User.destroy({ where: { id: testUser0.id } }),
-        db.User.destroy({ where: { id: testUser1.id } })
       )
+      if (testUser1) {
+        promises = _.concat(promises,
+          db.User.destroy({ where: { id: testUser1.id } })
+        )
+      }
+      if (testUni0) {
+        promises = _.concat(promises,
+          db.University.destroy({ where: { id: testUni0.id } })
+        )
+      }
 
       Promise.all(promises).then(() => {
         done()
       })
-    });
+    })
   })
 })
